requirements.txt
astroid==2.12.12
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
dill==0.3.6
Flask==2.2.2
idna==3.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.5
pyOpenSSL==22.1.0
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
